plugins/ml-training/skills/example-projects/examples/text-generation/requirements.txt
torch>=2.0.0
transformers>=4.30.0
peft>=0.5.0
datasets>=2.14.0
pyyaml>=6.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
